# src/main/resources/static/swagger/mng-api-docs.yaml
openapi: 3.0.3
info:
  title: 관리자 REST API 명세 ( yaml )
  description: 관리자 REST API 명세 with Swagger ( yaml )
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - Authorization: []
paths:
  /api/mng/user/list:
    get:
      tags:
        - 관리자 회원 컨트롤러
      summary: 회원 목록
      description: 회원 목록 조회
      operationId: userList
      parameters:
        - name: userVO
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/사용자 DTO'

      responses:
        "200":
          description: OK
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/ResultVO'
              examples:
                test:
                  value:
                    searchKeyword: "2222"
                    pageIndex: "1"
                guide:
                  value:
                    searchKeyword: "회원 PWD"
                    pageIndex: "회원 순번"
  /api/mng/user/proc:
    put:
      tags:
        - 관리자 회원 컨트롤러
      summary: 회원 수정
      description: 회원 수정
      operationId: userUpdateProc
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/사용자 DTO'
            examples:
              test:
                value:
                  user_seq: "5"
                  user_pwd: "2222"
                  user_auth: "ROLE_USER"
                  user_name: "테스트"
                  user_email: "test@test.com"
              guide:
                value:
                  user_seq: "회원 순번"
                  user_pwd: "회원 PWD"
                  user_auth: "회원 권한"
                  user_name: "회원 이름"
                  user_email: "회원 이메일"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/ResultVO'
    post:
      tags:
        - 관리자 회원 컨트롤러
      summary: 회원 등록
      description: 회원 등록
      operationId: userSetProc
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/사용자 DTO'
            examples:
              test:
                value:
                  user_id: "test_mng001"
                  user_pwd: "2222"
                  user_auth: "ROLE_USER"
                  user_name: "테스트"
                  user_email: "test@test.com"
              guide:
                value:
                  user_id: "회원 ID"
                  user_pwd: "회원 PWD"
                  user_auth: "회원 권한"
                  user_name: "회원 이름"
                  user_email: "회원 이메일"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/ResultVO'
    delete:
      tags:
        - 관리자 회원 컨트롤러
      summary: 회원 탈퇴 및 복구
      description: out_yn 값으로 회원 탈퇴 및 복구
      operationId: userDeleteProc
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/사용자 DTO'
            examples:
              test:
                value:
                  user_seq: "24"
                  flag: "d"
                  out_yn: "Y"
              guide:
                value:
                  user_seq: "회원 순번"
                  flag: "플래그 (d) 이미 탈퇴면 복구"
                  out_yn: "탈퇴 여부 (Y or N)"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/ResultVO'
  /api/mng/user/check:
    post:
      tags:
        - 관리자 회원 컨트롤러
      summary: 회원 ID 중복 체크
      description: 회원 아이디 중복체크
      operationId: getUserCheck
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/사용자 DTO'
            examples:
              test:
                value:
                  user_id: "test_mng001"
              guide:
                value:
                  user_id: "회원 ID"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/ResultVO'
...
  /mng/user/check:
    get:
      tags:
        - user-mng-controller
      operationId: getUserCheck_1
      parameters:
        - name: userVO
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/사용자 DTO'
      responses:
        "200":
          description: OK
          content:
            application/json;charset=UTF-8:
              schema:
                type: boolean
    put:
      tags:
        - user-mng-controller
      operationId: getUserCheck_4
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/사용자 DTO'
      responses:
        "200":
          description: OK
          content:
            application/json;charset=UTF-8:
              schema:
                type: boolean
    post:
      tags:
        - user-mng-controller
      operationId: getUserCheck_3
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/사용자 DTO'
      responses:
        "200":
          description: OK
          content:
            application/json;charset=UTF-8:
              schema:
                type: boolean
    delete:
      tags:
        - user-mng-controller
      operationId: getUserCheck_6
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/사용자 DTO'
      responses:
        "200":
          description: OK
          content:
            application/json;charset=UTF-8:
              schema:
                type: boolean
    options:
      tags:
        - user-mng-controller
      operationId: getUserCheck_7
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/사용자 DTO'
      responses:
        "200":
          description: OK
          content:
            application/json;charset=UTF-8:
              schema:
                type: boolean
    head:
      tags:
        - user-mng-controller
      operationId: getUserCheck_2
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/사용자 DTO'
      responses:
        "200":
          description: OK
          content:
            application/json;charset=UTF-8:
              schema:
                type: boolean
    patch:
      tags:
        - user-mng-controller
      operationId: getUserCheck_5
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/사용자 DTO'
      responses:
        "200":
          description: OK
          content:
            application/json;charset=UTF-8:
              schema:
                type: boolean
...
  /api/mng/banner/detail/{banner_seq}:
    get:
      tags:
        - 배너 관리 컨트롤러
      summary: 배너 상세정보
      operationId: bannerDetail_1
      parameters:
        - name: banner_seq
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/ResultVO'
...
components:
  schemas:
...
    ResultVO:
      type: object
      properties:
        resultMsg:
          type: string
          description: 응답 메시지
        resultCode:
          type: integer
          description: 응답 코드
          format: int32
        result:
          type: object
          additionalProperties:
            type: object
            description: 응답 결과 데이터
          description: 응답 결과 데이터
      description: REST API 응답 결과 객체
...
  securitySchemes:
    Authorization:
      type: http
      scheme: bearer
      bearerFormat: JWT